extracted Fee into a type alias

+++ txsystem/money/tx_executor_reclaim_fee_credit.go
@@ -55,7 +55,7 @@ func handleReclaimFeeCreditTx(s *state.State, hashAlgorithm crypto.Hash, trustBa
 			if !ok {
 				return nil, fmt.Errorf("unit %v does not contain bill data", unitID)
 			}
-			newBillData.V += v
+			newBillData.V += uint64(v)
 			newBillData.T = currentBlockNumber
 			newBillData.Backlink = tx.Hash(hashAlgorithm)
 			newBillData.Locked = 0
